+++ src/crystal/server.py
@@ -31,6 +31,7 @@ import json
 import re
 import socket
 import socketserver
+import sys
 from textwrap import dedent
 import time
 from typing import Literal, Optional, TYPE_CHECKING
@@ -109,6 +110,16 @@ class ProjectServer:
                         raise
                 else:
                     break
+            
+            if port == _DEFAULT_SERVER_PORT and tests_are_running():
+                print(
+                    '*** Default port for project server is in use. '
+                    'Is a real Crystal app running in the background? '
+                    'Will continue with a different open port, '
+                    'but the current automated test may not expect '
+                    'the server to be running on a non-default port.',
+                    file=sys.stderr
+                )
             
             # Try another port
             port += 1